+++ scripts/basic.py
@@ -4,4 +4,3 @@ from compas.geometry import Sphere
 viewer = Viewer()
 viewer.add(Sphere([0, 0, 0], 1))
 viewer.run()
-