+++ src/compas_viewer/components/renderer/__init__.py
@@ -1,1 +1,0 @@
-# from .render import Render  # noqa: F401