+++ src/compas_viewer/components/renderer/objects/__init__.py
@@ -1,1 +1,0 @@
-from .object import ViewerObject  # noqa: F401