+++ CHANGELOG.md
@@ -8,7 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 
 ### Added
-* Basic `Viewer` buildup.
+
+* Basic `Viewer` buildup with the widgets.
 * Configurations: `controller_config` and `viewer_config`
 * Code and File structure diagrams.